ref: modified returned value from float to double

+++ include/Engine/Core/Timer.hpp
@@ -11,7 +11,7 @@ namespace eng::core
 		Timer();
 
 		void restart();
-		float getElapsedSeconds();
+		double getElapsedSeconds();
 
 	protected:
 		std::chrono::steady_clock::time_point m_start_time;
